refactor(viewer): dissolve FeatureBar into FeatureItem and viewer store
- remove dead FeatureBar store fields (width, listStyle, showCellOutline)
  superseded by the generic Sidebar setup and the viewer store
- remove unreachable isExpanded channel filter and its orphaned
  ToggleExpander toggle; channel list now always shows all channels
- move pixelValues into the viewer store as a flat, non-persisted field;
  previously persisted to localStorage on every hover event
- rename FeatureBar/ to FeatureItem/ (generic collapsible wrapper plus
  per-item store factory), move Presets to its own directory
