+++ app/components/.client/ImageViewer/components/ChannelsController/ChannelsController.tsx
@@ -4,13 +4,10 @@ import { ChannelsControllerItemList } from "./ChannelsControllerItemList";
 import { Histogram } from "./Histogram";
 import { select } from "../../state/store/selectors";
 import { useViewerStore } from "../../state/store/ViewerStoreContext";
-import { FeatureItem } from "../FeatureBar/FeatureItem";
-import { useFeatureBarStore } from "../FeatureBar/useFeatureBar";
+import { FeatureItem } from "../FeatureItem/FeatureItem";
 
 export function ChannelsController() {
   const layersStates = useViewerStore(select.layersStates);
-  const isExpanded = useFeatureBarStore((state) => state.isExpanded);
-  const setIsExpanded = useFeatureBarStore((state) => state.setIsExpanded);
   const channelsOpacity = useViewerStore(select.channelsOpacity);
   const setChannelsOpacity = useViewerStore(select.setChannelsOpacity);
   const visibleChannelCount = useViewerStore(select.visibleChannelCount);
@@ -40,7 +37,7 @@ export function ChannelsController() {
     >
       {layersStates.map((_, index) => (
         <TabPanel key={index} id={String(index)}>
-          <ChannelsControllerItemList isExpanded={isExpanded} setIsExpanded={setIsExpanded} />
+          <ChannelsControllerItemList />
         </TabPanel>
       ))}
     </FeatureItem>
