+++ app/components/.client/ImageViewer/state/store/__tests__/createViewerStore.test.ts
@@ -62,9 +62,11 @@ describe("createViewerStore", () => {
       imagePanelIndex: -1,
       imagePanels: [],
       cursorPosition: null,
+      pixelValues: {},
       layersStates: [],
       setError: expect.any(Function),
       setCursorPosition: expect.any(Function),
+      setPixelValues: expect.any(Function),
       setViewStatePreview: expect.any(Function),
       setViewStateActive: expect.any(Function),
       setIsViewerLoading: expect.any(Function),
